third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-frame-move-expected.txt

Verifies mouse event co-ordinates from a tap in an iframe that changes position on each mouse event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Sending GestureTapDown to 25,25
Sending GestureShowPress
Sending GestureTap


Received mousemove in child frame with currentOffset=0 expectedOffset=0
PASS event.screenX is 25
PASS event.screenY is 25
PASS event.clientX is 25
PASS event.clientY is 25
PASS event.pageX is 25
PASS event.pageY is 25
PASS event.target.nodeName is "DIV"
Target element is at (25,25,30,30)
PASS targetRect.right is >= event.clientX
PASS event.clientX is >= targetRect.left
PASS targetRect.bottom is >= event.clientY
PASS event.clientY is >= targetRect.top


Received mousedown in child frame with currentOffset=5 expectedOffset=5
PASS event.screenX is 25
PASS event.screenY is 25
PASS event.clientX is 20
PASS event.clientY is 20
PASS event.pageX is 20
PASS event.pageY is 20
PASS event.target.nodeName is "DIV"
Target element is at (20,20,25,25)
PASS targetRect.right is >= event.clientX
PASS event.clientX is >= targetRect.left
PASS targetRect.bottom is >= event.clientY
PASS event.clientY is >= targetRect.top


Received mouseup in child frame with currentOffset=10 expectedOffset=10
PASS event.screenX is 25
PASS event.screenY is 25
PASS event.clientX is 15
PASS event.clientY is 15
PASS event.pageX is 15
PASS event.pageY is 15
PASS event.target.nodeName is "DIV"
Target element is at (15,15,20,20)
PASS targetRect.right is >= event.clientX
PASS event.clientX is >= targetRect.left
PASS targetRect.bottom is >= event.clientY
PASS event.clientY is >= targetRect.top


Received click in child frame with currentOffset=15 expectedOffset=10
PASS event.screenX is 25
PASS event.screenY is 25
PASS event.clientX is 15
PASS event.clientY is 15
PASS event.pageX is 15
PASS event.pageY is 15
PASS event.target.nodeName is "BODY"
Target element is at (0,0,500,500)
PASS targetRect.right is >= event.clientX
PASS event.clientX is >= targetRect.left
PASS targetRect.bottom is >= event.clientY
PASS event.clientY is >= targetRect.top
PASS eventReceived.mousemove is true
PASS eventReceived.mousedown is true
PASS eventReceived.mouseup is true
PASS eventReceived.click is true
PASS successfullyParsed is true

TEST COMPLETE
